Разносит константы по отдельным файлам
