+++ src/components/page-nav/page-nav.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { AppRoute } from '../../const';
+import { AppRoute } from '../../const/server';
 
 function PageNav(): JSX.Element {
   return (
