+++ src/components/private-route/private-route.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
+import { AppRoute, AuthorizationStatus } from '../../const/server';
 
 type PrivateRouteProps = {
   authorizationStatus: typeof AuthorizationStatus[keyof typeof AuthorizationStatus];
