+++ src/const/cities.ts
@@ -1,0 +1,10 @@
+const Cities = [
+  'Paris',
+  'Cologne',
+  'Brussels',
+  'Amsterdam',
+  'Hamburg',
+  'Dusseldorf',
+] as const;
+
+export { Cities };
