+++ src/const/modes.ts
@@ -1,0 +1,11 @@
+const CardMode = {
+  Default: 'default',
+  Favorite: 'favorite',
+} as const;
+
+const LogoMode = {
+  Header: 'header',
+  Footer: 'footer',
+} as const;
+
+export { CardMode, LogoMode };
