+++ src/pages/main-page/main-page.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import PageHeader from '../../components/page-header/page-header';
 import LocationsTabs from '../../components/locations-tabs/locations-tabs';
 import OffersList from '../../components/offers-list/offers-list';
-import { Offer } from '../../types/offer';
-import { Cities } from '../../const';
+import { Offer } from '../../types/types';
+import { Cities } from '../../const/cities';
 
 type MainPageProps = {
   offers: Offer[];
